moves DynatraceDashboard (and friends) to own file and converts function to method
